requirements-snapshot.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
autopage==0.6.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
celluloid==0.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cliff==4.13.3
cmaes==0.13.0
cmd2==3.5.0
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
cypari==2.5.6
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
einops==0.7.0
et-xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.28.0
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
fxrays==1.3.6
gitdb==4.0.12
gitpython==3.1.46
gnureadline==8.3.3
graph-universe==0.1.2
grpcio==1.80.0
gudhi==3.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
hypernetx==1.2.5
idna==3.11
igraph==1.0.0
ipykernel==7.2.0
ipython==9.10.1
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
kiwisolver==1.5.0
knot-floer-homology==1.2.2
lark==1.3.1
lightning==2.4.0
lightning-utilities==0.15.3
littleutils==0.2.4
low-index==1.2.1
mako==1.3.11
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
notebook==7.5.5
notebook-shim==0.2.4
numpy==1.26.4
ogb==1.3.6
omegaconf==2.3.0
openpyxl==3.1.5
optuna==2.10.1
outdated==0.2.2
overrides==7.7.0
packaging==24.2
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.2.0
platformdirs==4.9.6
plink==2.4.9
prettytable==3.17.0
prometheus-client==0.25.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.13.2
pydantic-core==2.46.2
pyg-nightly==2.8.0.dev20260418
pygments==2.20.0
pyparsing==3.3.2
pyperclip==1.11.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-igraph==1.0.0
python-json-logger==4.1.0
pytorch-lightning==2.6.1
pyx==0.17
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==15.0.0
rich-argparse==1.7.2
rootutils==1.0.7
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
send2trash==2.1.0
sentry-sdk==2.58.0
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
snappy==3.3.2
snappy-manifolds==1.4
soupsieve==2.8.3
spharapy==1.3.0
spherogram==2.4.1
sqlalchemy==2.0.49
stack-data==0.6.3
stevedore==5.7.0
sympy==1.14.0
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tkinter-gl==1.1
topobench @ file:///Users/cretuluca/uni/TopoBench
topomodelx @ git+https://github.com/pyt-team/TopoModelX.git@8fc047c85886642cf53a784fc712cbae89b015fd
toponetx @ git+https://github.com/pyt-team/TopoNetX.git@c378925c52169fd46bf81868dc6bf7c18f81a0bc
torch==2.3.0
torch-cluster==1.6.3
torch-geometric==2.7.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
torchmetrics==1.9.0
torchvision==0.18.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
trimesh==4.11.5
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
uri-template==1.3.0
urllib3==2.6.3
wandb==0.26.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
werkzeug==3.1.8
wheel==0.46.3
xlrd==2.0.2
xxhash==3.6.0
yarl==1.23.0
